feat(adapter): getUsage now syncs if necessary
